move store into its own top level mod
